ENH: handle enstaller in the requirement list.
If enstaller is required as is, we've already made sure before that
enstaller is up to date, so we just mention it and remove it from the
list of reqs.

If enstaller is required with a specific version, it can only be a
downgrade at that point, so we warn about it and continue.